test: Split __tests__/redirects.ts into two files
